fix interface orientation portrait

+++ AKN/TablePageViewController.m
@@ -128,7 +128,7 @@
 {
     UICollectionViewCell *cell=[collectionViewNews dequeueReusableCellWithReuseIdentifier:@"cell1" forIndexPath:indexPath];
     UIImageView *img=(UIImageView*)[cell viewWithTag:20];
-    img.image=[UIImage imageNamed:[NSString stringWithFormat:@"%d.jpg",(indexPath.row+1)]];
+    img.image=[UIImage imageNamed:[NSString stringWithFormat:@"%ld.jpg",(long)(indexPath.row+1)]];
     return cell;
 }
